requirements.txt
Flask
openai
python-dotenv
SQLAlchemy
psycopg2-binary
